pyproject.toml
[tool.poetry]
name = "shibudo finances - api"
version = "0.1.0"
description = "Projeto para controlar suas finanças de forma fácil e intuitiva"
authors = ["Gabriel Machado <gbrasil200@gmail.com>"]
readme = "README.md"
packages = [{include = "shibudo finances _ api"}]

[tool.poetry.dependencies]
python = "^3.11"
djangorestframework-simplejwt = "^5.2.2"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.5"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"
ipdb = "^0.13.11"
pycodestyle = "^2.10.0"
ipython = "^8.11.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
